refactor(#57): pr 수정
새로고침시 자동 로그인이 풀리는 문제 해결
